restructure and rename some folders
