+++ src/api/models/NewsModel.ts
